Restore slightly tweaked RawMemoryStream to fix bug #2961787: "HE SAM1: Music in kitchen slightly off (regression)".
The mentioned tweak is only the removal of the templated "stereo" parameter
like it was done for RawStream, thus that is totally safe and saves some
binary size. Note that I was unable to test whether the HE regression was
really fixed, since I do not have any of the affected games and the Moonbase
Commander demo does not work for me. I tried the RawMemoryStream code with
KYRA and it worked fine.

svn-id: r48506