TOOLS: Fix MSVC 2010 project files.
Formerly the MSVC 2010 project files did not respect the global include/library/executable directories of the user, but used custom paths for that. This resulted in headers and libraries not be found by the MSVC, in case they were not stored in %(SCUMMVM_LIBS).

Additionally it overwrote the executable MSVC used when compiling for x64, thus breaking the x64 target.

svn-id: r51213