OPENGL: Properly initialize the graphics manager again.
Starting with r52503 the graphics manager's only initialize their event
observer when initEventObserver is called. It seems like it was an
oversight that this was not done in OSystem_SDL::setGraphicsMode, when a new
graphics manager was initialized.

This should fix window resizing with the OpenGL graphic's manager and mouse
movement.

svn-id: r52506