VIDEO: Mirror QuickTime's mishandling of its own mdat atoms
While the specs say that the 'mdat' ("media data") atom is required in all QuickTime files, QuickTime makes no check if the atom is actually present. This allows for some videos with typos in the atom tag in them to play in QuickTime anyway. The offsets in the 'stco' atom are only used instead.

This will fix a video in DS2 Mac and a music file in IHNM Mac (although this is not needed for either just yet).

svn-id: r55125